Código fuente linuxizado.

+++ Nucleo/CString.h
@@ -180,8 +180,8 @@ public:
 
     void Split ( std::vector < CString >& dest, char cSeparator = ' ', unsigned int uiMax = (unsigned int)-1 ) const
     {
-        int iPos = 0;
-        int iPos2 = 0;
+        size_t iPos = 0;
+        size_t iPos2 = 0;
         while ( ( iPos = find ( cSeparator, iPos ) ) != npos )
         {
             if ( static_cast < unsigned int > ( iPos ) > uiMax )
